backend/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.2
python-multipart
python-dotenv
email-validator
